Turn ARC On for 64bit compiles, don't make MIDI, fix various issues with bridge and windowIndex for ARC, fix parseArgument data type warning
git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3697 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 292bd22b6803d87398f2e3a95da322134c528c91

+++ platforms/iOS/vm/OSX/sqSqueakOSXApplication.h
@@ -45,7 +45,7 @@
 }
 @property (nonatomic,strong) NSCursor	*squeakCursor;
 
-- (NSInteger) parseArgument: (NSString *) argData peek: (char *) peek;
+- (int) parseArgument: (NSString *) argData peek: (char *) peek;
 - (void) parseArgs: (NSArray *) args;
 - (void) parseEnv: (NSDictionary *) env;
 - (long long) strtobkm: (const char *) chr;
